New Iteration of KlamathOps.mdl.gz
This version is able to tranfer flows from Williamson River down through Iron Gate Dam for the period of Feb 28 to Oct 16, 2018. Of the changes made, a few of the initialization rule were updated to use the altered objects, object slots set with values from the IGD Calc., and new historical records added to the data objects.