IGD Outflows have been corrected and the lag between UKL and IGD incorporated.
The 7 day lag has been assigned to  the UKL to Keno Routing object to mimic the models operations. To allow the model to succesfully run, pool elevation at IGD was allowed to fluctuate and the unregulated spill method turned off. Rules 107 and 121 were edited so to provide the correct inputs and thus outputs for rule 106 (IGD Outflow Projections.)