Edited the Rules and Functions to get the model running.
Edited or Deleted Functions in the “Model Functions” Group due to their dependent slots being deleted or transferred when the “Reservoir Data” and “KBPMOperationsData” Data Objects were deleted.
Repeated the same process for the “Operation Functions” Group but decided to disable it due to the time it was taking to complete the edits and it not being utilized for any of the rules we had written.
Disabled the “Lost River Operations” Rule Group as it was not needed for the operations I was doing on the Klamath River, these will likely need to be debugged before being valid.
Edits to Objects:
-	Deleted A1 Northwestern Return Colum in Return Flow Propagation Slot
-	Set the Available Flow Method to “Available Flow Diversion” for both the Ady Canal Diversion and Center Canal Diversion Objects.