New Policy Sets Added to the Model
Created the Summer Operations Policy Set containing the Following Rules
-	Compute UKL for River
-	Set IGD Outflow
-	Compute UKL Corrected Override Flow
-	Compute UKL Ramping & Proj Min Release
-	Compute UKL for Flood Control
-	Compute IGD Corrected Override Flow
-	Compute IGD Preliminary Flow
-	Compute UKL for IGD
-	Compute IGD Summer Ramping
-	Compute Spring Fill Rate
-	Compute UKL Max Elevation
-	Compute Williamson Cumulative Proportion
-	Compute Williamson Cumulative Inflow
-	Compute EWA Remain
-	Compute EWA Used
-	Compute EWA Reserve
-	Compute EWA & Will50
-	Compute F AND FF Pump Outflow Reduction
Created the Dilution and Flushing Policy Set containing the Following Rules
-	Compute Cumulative Dilution Flow
-	Compute Long Dilution Flow
-	Compute Short Dilution Flow
-	Compute Surface or Deep Flushing
-	Compute IGD Release for D&F
-	Compute Day Count
-	Compute IGD with Accrete
Edits to existing rules were made to account for summer operations as well as handle the change in starting timestep to Feb 22nd
In addition, numerous slots and functions were created to run the above rules. These will not be listed as the list is rather long.
Currently, the rules operate for the entire run length. The values, however, only match the IGD Calc until July 1, 2018 as this is as far as I’ve debugged.