Tested Model with Operation Start on Dec 1, 2018
Minor changes to rules were made during debugging to better match the Model outputs with the IGD Calc. In addition, many hard input constants or DateTimes were changed to be referenced to slots on objects to ease future editing.
An R Script was developed to automate the comparison between the IGD Calc. and the RiverWare Model. Instructions on how to complete the testing have been uploaded as well.