Further changes to KROM_Export tab of the backup sheet
I reorganized the data again and added a column for Excel date and one for RiverWare date to make it clear why it is needed. With this new column, I had to change the DMI to offset the columns by one. That is why the model changed too.

I also reorganized the columns to put the input column at the end so it isn't filled down.