Updated PA Calc with changes made to Ady Canal by KBAO
Due to some copy paste error, the PA Calc did not have the correct data in for Ady Canal.  KBAO fixed it in version 29 as follows by Baker: "made changes to Ady Canal, Ady to LNKWR, and Ady to Ag columns in the PA Calculator that should now be consistent with the Daily Data Import (Backup) sheet."

Since we are using version 28 for testing, I applied those same changes to a version 28 and saved it as *Rev28_DN_20200909.xlsx.  This is the version that should be used for testing.