Updated Scripts and Fixed an Modeling error
This commit fixes a serious issue having to do with the inflows. When moving UKL Inflow computations to the object dispatching, we neglected to notice that there was an Init rule computed the forecast volume that referenced the UKL Inflow. This Init rule failed and didn't set a value. Whatever was there from previous single or MRM runs would be used. Not good. Now we compute the Inflow on the Inflow Init slot on UKL using an Init Rule. This fixes it.

Also, improved the scripts for usability and clarity.