Added functionality to allow IGD Outflow to be input
Based on comments at the workshop, we need a good way to override IGD Outflow with an input.  This was possible, but I screwed up the demo.
Now it is much clearer and you can input it directly on the SCT and the rules/functions recognize the input and it doesn't abort.