Work in progress - Conversion of Dashboard.SupplyConsiderations table into series slots
This commit moves the Dashboard.SupplyConsidrerations into series slots on the Dashboard object. This is necessary as some of the columns, EWA Augment, and Project Supply Reduction, are now computed at the start of March (Start of April should be added later).  The initialization rules were modified to compute all of the values in this table at the start of the run. I added rules IR 8, 9, 12, 13, 14 and modified a bunch of other init and RBS rules to point to the new slots.

I tried to use RBS rules but some of the computed variables are dependent on other tables that are computed at initialization.  These will all be moved to RBS rules as we get to planning mode in earnest.

Limitations:
- April Lock In is a stub using test data.  It can be input though.  We should add it to the SCT then.
- The init rules have many literal numbers. They need to be cleaned up
- I didn't test this in MRM for OM. It likely doesn't work
- I didn't test fully for a range of ops start dates.

I also updated the script dashboard to give correct dates on valid Ops Start Timesteps.