Fixed up Supply Init rules
Init rule 14 sets April Lock In and Switch. It is a stub using test data.  It can be input if there is real data, so I added it to the SCT Per Run Settings tab.

Init rule 9 (Set EWA Augment Volume...) had many literal numbers. I added two slots to hold the thresholds and volumes.  Also, I reworked the rule so it sets the augment volume using the supply at the start of each month in consideration.

Limitations:
- I didn't test this in MRM for OM. It likely doesn't work
- I didn't test fully for a range of ops start dates.