Moved Supply rules from Init set to RBS Set
This is a preliminary working version of the model with all of the rules that set Project, UKL and EWA Supplies moved from the Initialization set to the RBS set where they are now setting series slot values.

One key to this is that a very low priority rule "Set Forecast Timestep Flag" decides if it is a "Forecast Timestep" by looking for a manual input on the "NRCS Forecast Table" or the "Inflow Forecast".  If it is a forecast timestep, a 1 is set on the "Season and Dates.Iron Iron Gate Target Date" slot. Otherwise, nothing is set.  We'll modify this in planning mode to decide when it is a forecast timestep, i.e. Mar 1, Apr 1, etc.

Many rules then key off of the "Season and Dates.Iron Gate Target Date"  slot to decide if they should be recomputed or just use the previous timestep value.

Most of these changes are in the low priority groups, "Account Supply" and "Forecasts".

The model runs and produces reasonable UKL elevations and flows.  The next step will be further testing and debugging to make sure everything is working.  Once satisfied, then we can delete the disabled rules and assignments. Then move on to planning mode.