Moved Demand Forecasts from Init Rules to RBS rules
This commit moves the rules in the Demand Forecasts group from the Init Rules to the RBS rules. Then they are edited to set values each timestep instead of looping over the run.   I duplicate and disabled a version of the logic so we can see what it said. Eventually, we should delete the disabled rules and assignments.

I also removed the Ag Initialization Init Rule as all it did was set a series slot from a periodic slot. This is not needed as the rules can reference the periodic slot directly.