More minor changes based on testing
Added logic to "Set Forecast Series Slots" to set the forecast to the "Alt UKL Net Inflow Forecast"  if we don't yet have an NRCS forecast. This is only valid during March if there isn't a previous value.  I did a similar thing for the Forecast Adj. It uses 1.0 if not input yet.

Fixed a script error due to a changed slot name.

I tested for various ops start dates between Oct 2020 and Sep 30 2021 and it seems to be running fine.  I didn't look into results though.