Added Slots and DMI to bring in Historical NRCS UKL Inflow forecast. Set up a function for Planning Mode
I added the slot: Historical NRCS UKL Water Supply Forecast and a DMI to bring it in from the Archive spreadsheet.
The name of the slot in the spreadsheet now matches the name in the model.

Also renamed slots on the Forecasts object to say UKL before Inflow on most slots.  Rules and other RPL logic were updated as appropriate.
I changed the slot Inflow Forecast to "UKL Inflow Forecast for Run". This is the slot that is set in various ways based on the mode but it is the one used in the run.

Finally, I added a function Is Planning Mode and some logic in rule 33 to start to use the mode.  Now in Planning mode, the "Forecasts.UKL Inflow Forecast for Run" is set to the Historical NRCS UKL Water Supply Forecast.