Made minor changes for Planning mode
Fixed the DMI and spreadsheet so both volume columns have "acre-ft" units.

Added "Prepare Planning Run" script.

Made changes to the ruleset to flag when a forecast changes in planning mode. This is a forecast date.
Now it is not based on a non NaN in the Seasons and Dates.Iron Gate Target Date; it must be 1.0 to be a forecast date.