Deleted all disabled rules and assignments
I Deleted all disabled rules and assignments in the RBS, MRM, and Init Sets

There are still many slots that are marked as DELETE or REMOVE that need to be deleted.

I also cleaned up the object positions so the Metric Canvas can be shown on the workspace.

Finally, I made a new output canvas to show Run Information on the workspace.