Set forecasts and Accretions for 1 WY in planning mode
This commit adds in the rules and functions to set Inflows and Accretions for only 1 WY at a time when in planning mode.

I also fixed some errors in the stubbed out logic that finds the distribution percentage based on time of year.

There are still lots of TODOs in comments, but it is running planning mode from Oct 1 2020 to Feb 28, 2022 and producing reasonable results

For the SCT, it is now saved with the setting enabled to show Priority values.