Change to add in Max Release table data, Max Out function and a new plot
I added two rows to UKL.Max Release Table to complete the lower end. The Elevation Volume Table starts at 4136, so this table should too. Previously, it started at 4137.
4136    0
4136.99    0

I also added a user defined MaxOut function (calls GetMaxOutflowGivenInflow) and applied it to rules 2 (refactored) and 24 that was getting the "No Spillways available" message.

Finally, I added a plot that shows modeled vs historic PE. This won't have a particularly good comparison as the policies that are modeled are different, but it provides general trends of wet vs dry.