Improvements when switching modes and a plotting change
First, I removed a synchronize slots action in the Setup Model for new WY script.  The slots are no longer series so don't need to be synced.

Also, I changed the Observed and Foretasted Elevation slots that are used for plotting to evaluate the expression at end of TS instead of end of run. This helps in long or failed runs so that something can be plotted.

WIth this change, it will round trip from Ops to Planning and then back to Ops using the appropriate scripts.