KPOM Model and input Spreadsheet - Adjusted number of timesteps for some variables
In the model, I modified the DMI to only bring in as much prerun data as necessary.  This shortened the imported data sereis for a few slots:
Ady to Refuge.Diversion Request - (Was 7 timesteps; Change to -1 timesteps Sep 30)
Ady Canal.To Ag - (Was 7 timesteps; Changed to 1 timesteps Sep 30)
11509500 Klamath at Keno.Gage Inflow - (was 60 timesteps; Changed to 7 timesteps Sep 24)
11516530 Klamath bl Iron Gate.Gage Inflow - (was 60 timesteps; Changed to 1 timesteps Sep 30)
Ady Canal.Diversion Request - (was 60 timesteps; Changed to 7 timesteps Sep 24)
Miller Hill Pump.Diversion Request - (was 60 timesteps; Changed to 7 timesteps Sep 24)
North Canal.Diversion Request - (was 60 timesteps; Changed to 7 timesteps Sep 24)
Station 48.Diversion Request - (was 60 timesteps;  Changed to 7 timesteps Sep 24)
Other slots were not changed.

I also put in the spreadsheet the number of timesteps of pre-run data that is expected. This is in the first row.