Added logic to compute Hydrologic Year Type for Flood Curve
I added in rule 6 "Set Hydrologic Year Type for Flood Control".
In ops mode, it uses the input periodic slot.
In planning mode, it looks up the Water Supply forecast on Jan 1, Feb 1, Mar 1, and Apr 1 and sets it to Wet (3) or Dry (1).

As part of this, I modified the slots accessed by these rules.