Moved the official development model back to WY2020
This commit moves the official model back to WY2020, that is:
Run Range: 10/1/2019 to 2/28/2021
Ops Start Timestep: 7/22/2020

So why do this?
As we discussed, in operations mode, we should be able to duplicate the results from the PA Calculator and KROM that we made in 2020.  More specifically, we want to reproduce results from the PA Calculator V28 that we used for testing in previous efforts.  Moving the official model back to this range makes it easier to maintain that connection to v28 and the previous KROM.

It is still easy to move the model to a longer range for planning mode or move it forward to 2021 for additional ops mode testing, all using scripts.

In addition to moving the run range, I updated the operations data to use the values from KROM/PACalc28. This is my checklist:
- Copy in UKL.InflowExceedance monthly values
- Copy in Distribution Fractions from Testing/DistFractionsForTesting.2020.07.22.xlsx
- Edit Forecasts.NRCS Forecast Table - Edits dates to be in 2020 instead of 2021
- Check all diversion Exceedance and Adj Input slots
- Copy/Paste series values from Per Run Settings in SCT

After making all these changes, the results are close to KROM 2020 and PACalc28 but not exact.  Now the debugging can proceed.