KPOM Models, Rules, and GFS - UKL Release to River
I added a slot to UKL: UKL.Release to River
It is set by Rule 18 "Set Environmental UKL Release" and reset by Rule 1 "UKL Release for bl IGD Flows"
It is my understanding that the UKL Release for IGD falls into the category of "Release to River" and is charged to EWA.
To do this, I also reorganized rule 1 to use an IF and WITH statement.

I also reworked the function "UKL Release to River(date)".  If it is the current timestep, compute the value. If it is a past timestep, use the value on the new slot at that timestep.