Fixes to get KPOM running in Planning mode again
On 7/28, a commit was pushed where the Diversion Requests were added to the Historical Water Supply Forecast DMI. The log says: "modified DMI's and Initialization rules for planning runs"
This DMI is run by the planning mode setup script. The DMI locked in the Diversion Requests (Z flags) and prevented rules from cutting them back.  I removed that slot selection from the DMI as I don't know what it was there for.

An additional issue I found:  a commit on 8/11 of the Archive xlsx sheet threw away two columns that were previously in there. IGD.Historical Pool Elevation (ft)  and  Keno to IGD.Forecast Accretions (cfs).  I believe the intent was to remove only the second of those two columns.

I also found a divide by zero in an expr slot. I fixed that too. In the Demand from UKL global function, if the denominator is 0, the entire expr is zero.