Updated a few rules to get it to run in Ops Mode
This commit makes a few changes to run in operations mode to get UKL Inflow set:
- RBS rule "Compute Smoothed Inflows" to set UKL.Inflow during operaions period only.
- RBS rule "UKL Inflow Forecasts" to include timestep before Operations Start Timestep
- Init rule "UKL Inflow Forecast Exceedance - Ops Mode" - include timestep before Operations Start Timestep

Changed the name of the new WY script to say it is used for OM.

Reassigned a slot in the SCT to point to "EWA.EWA May June Augment Release".  I also put some data in that slot that was lost from the previous   EWA May June Augment slot for spring 2020 that was used in the test model.