Fixed rule 34 and added a per-year plotting prototype
I removed the execution constraint on rule 34,"Set Ag Adj Factor" as it stopped setting the "Supply.Agricultural Adj Factor" in Planning mode (to 1.0).  This prevented many of the PA Viewer expression slots to solve. It also prevented rule 14 from changing the diversion requests, which may be another issue.

I also added a prototype tool so the user can choose which years to plot when in Planning mode.   I duplicated 2 sample plots and then created and referenced a global function for the min and max dates for the new copied plots.  The function points to a scalar slot PAViewer.PlottingYear where you specify the WY.  Then I added  script "Create per-Year Plots" where the user can set the year and select the plots to generate.
I also added this year slot to the main sct too and am committing it.