Working commit of some model and function cleanup.
Changed a few functions and expressions that referenced 1 "year" as this will not be allowed in 8.6. To fix, I added in a DaysInYear function that checks for leap years and returns the correct value.

Changed the name of the ruleset to "Klamath 2019 BiOp and 2020 IOP"

Cleaned up some script descriptions

Added a script to run Planning Mode with Iterative MRM"PM: Make Planning Run using Iterative MRM"

Added script groups.