Additional changes for OM and MRM runs
I added some logic that was needed to get the OM MRM to work with the historical scenarios.
This was in the Init Rule and in RBS rule 53
Without these changes, the UKL inflows were not getting set.

I also update the 2 MRM plots to scale to run range and update each time.