New input DMIs for Ops mode and MRM demo
For the ops mode demo, we need to input a bunch of values for both MRM and single mode.
Instead of taking the time to do that, I set up a DMI that brings in the appropriate data.
The source is an excel sheet KPOM_OM_InputsForTesting.xlsx in the Database folder.