Changes for OM and PM - IGD targets weren't being met
This commit does two main things:
- recommits the changes I made yesterday at 10:36 am as these were lost:
"New input DMIs for Ops mode and MRM demo. For the ops mode demo, we need to input a bunch of values for both MRM and single mode.
Instead of taking the time to do that, I set up a DMI that brings in the appropriate data.
The source is an excel sheet KPOM_OM_InputsForTesting.xlsx in the Database folder.

- Reorders Rule 2 to become rule 4 and edits to correctly account for the lag.
It assumed a three day lag but we removed the physical lag.  Now we use the physical lag value.