Implemented scripts to perform Base and Alt as a prototype
This commit adds functionality to preserve a planning run as either the Base or Alt 1 using snapshots. Plots compare the two.  This was previously set up by Nathan to use Iterative MRM, but now that planning runs use Iterative MRM, we can't use it for Base/Alt.  Instead we have added:
- Two new scripts to "Preserve Results as Base or Alt" - This takes snapshots of key slots and then evaluates expression slots used for plotting
- Added a new Base and Alt Slots with expression slots that point to the simulation slots with results.  These expression slots are plotted.
- Two new snapshot objects Base and Alt 1
- Plots that mimic the PAViewer alt vs base plots.
- Slot sets to make the scripts more dynamic

In the future we can add more slots and plots to this and/or add Alt 2, 3, ...

I also deleted most of the plots names "Multiple Alternatives"
I changed the MRM rule names for the "Record slots for Base vs Alt Plots" and added a note to Base vs Alt MRM Dashboard that they are no longer used.
Finally, I changed a global function "IGD Storage Fcst" to use DATETIME as it was coded incorrectly.