Added script to help with plotting Base vs Alt
This is a follow on to the last commit. In this commit, I added a script (Create Plots of Base and Alt 1) that:
- evaluates the expression slots for Base and Alt 1 and
- generates the plots.
This is needed as the expr slot values aren't saved unless outputs are saved.
The Snapshot values are saved, so we just need to re-evaluate the expression slots and then create the plots.